feat: add modules tasks solution
